Support React root views which aren't children of content view
...on Android.

Previously only direct children of the content view would be checked for
being a `RootView`. This change allows any descendant of the content
view to be considered as the `RootView`.

This is useful for example when embedding React Native into an existing
native app. In cases like this, there may be native views which are
direct children of the content view, and the React Native root may be
further down the View hierarchy.

Without this change, when the React Native root was deeper in the View
hierarchy, it wasn't possible to send events to it via
`MSREventBridgeModule#emitEventForActivity`.